backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
httpx==0.25.1
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
